machine_learning/requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astor==0.8.1
astunparse==1.6.3
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
dlib==19.23.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.68.2
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h11==0.13.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.1
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
numpy==1.22.2
oauthlib==3.2.0
opt-einsum==3.3.0
Pillow==9.0.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.9.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
starlette==0.14.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.16.1
tensorflow-estimator==2.1.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tf2-yolov4==0.1.0
toml==0.10.2
typeguard==2.13.3
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.15.0
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
